% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{decode_4}
\alias{decode_4}
\title{Given a 4 digit code, retrieve the bounding coordinates
or center of a One Degree Square. One Degree Squares (ODS) are defined degree
by latitude and degree longitude}
\usage{
decode_4(x = 4269, quad = c(lat = 1, lon = 1), form = "bounds",
  closed = TRUE)
}
\arguments{
\item{x}{numeric or character one or more 4 digit ODS codes}

\item{form}{character either 'bounds' or 'center'}

\item{closed}{logical if form is 'bounds' then close the bounds}

\item{quadrisphere}{character identifier where 'nw' is northern western quadrisphere}

\item{a}{list of 2xn matrices where n = 1 for 'center' and n = 4 or 5 for 'bounds'}
}
\description{
Given a 4 digit code, retrieve the bounding coordinates
or center of a One Degree Square. One Degree Squares (ODS) are defined degree
by latitude and degree longitude
}

