% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{decode_5}
\alias{decode_5}
\title{Given a 5 digit code, retrieve the bounding coordinates
or center of a Quarter Degree Square. Quarter Degree Squares (QDS) use degree
Latitude and degree Longitude and quarter of the full degree square.
The lower right quarter is 1, lower left 2, upper right 3 and upper left 4.}
\usage{
decode_5(x = 42692, quad = c(lat = 1, lon = 1), form = "bounds",
  closed = TRUE)
}
\arguments{
\item{x}{numeric or character one or more 5 digit QDS codes}

\item{form}{character either 'bounds' or 'center'}

\item{closed}{logical if form is 'bounds' then close the bounds}

\item{quadrisphere}{character identifier where 'nw' is northern western quadrisphere}

\item{a}{list of 2xn matrices where n = 1 for 'center' and n = 4 or 5 for 'bounds'}
}
\description{
Given a 5 digit code, retrieve the bounding coordinates
or center of a Quarter Degree Square. Quarter Degree Squares (QDS) use degree
Latitude and degree Longitude and quarter of the full degree square.
The lower right quarter is 1, lower left 2, upper right 3 and upper left 4.
}

