% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{decode_id}
\alias{decode_id}
\title{Given a 4, 5 or 6 digit code, retrieve the bounding coordinates
or center of a fish box.}
\usage{
decode_id(x = c(426916, 42692, 4269), form = "bounds", closed = TRUE,
  quadrisphere = "nw")
}
\arguments{
\item{x}{numeric or character one or more six digit TMS codes}

\item{form}{character either 'bounds' or 'center'}

\item{closed}{logical if form is 'bounds' then close the bounds}

\item{quadrisphere}{character identifier where 'nw' is northern western quadrisphere}

\item{a}{list of 2xn matrices where n = 1 for 'center' and n = 4 or 5 for 'bounds'}
}
\description{
Given a 4, 5 or 6 digit code, retrieve the bounding coordinates
or center of a fish box.
}

