% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{decode_polygons}
\alias{decode_polygons}
\title{Retrieve Polygons class objects derived from 4, 5 or 6 digit codes used to
define One Degree, Quarter Degree or Ten Minute Squares (ODS, QDS, TMS).}
\usage{
decode_polygons(x = c(426916, 42692, 4269, 387066, 38701, 3870),
  crs = sp::CRS("+proj=longlat +datum=WGS84"), asDataFrame = TRUE, ...)
}
\arguments{
\item{x}{one or more numeric or character 4, 5 or 6 digit codes}

\item{crs}{a CRS class coordinate reference system object}

\item{asDataFrame}{logical, if TRUE return a SpatialPolygonsDataFrame}

\item{...}{further arguments for \code{decode_id}}
}
\value{
sp::SpatialPolygons or sp::SpatialPolygonsDataFrame class object
}
\description{
Retrieve Polygons class objects derived from 4, 5 or 6 digit codes used to
define One Degree, Quarter Degree or Ten Minute Squares (ODS, QDS, TMS).
}

