% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nefsc.R
\name{download_nefsc_gis}
\alias{download_nefsc_gis}
\title{Retreive one or more shape files from NMFS's NEFSC public GIS products}
\usage{
download_nefsc_gis(what = "Statistical_Areas", dest = ".",
  ftp_url = "ftp://ftp.nefsc.noaa.gov/pub/gis", exts = c("dbf", "prj",
  "sbn", "sbx", "shp", "shp.xml", "shx"))
}
\arguments{
\item{what}{one or more names of ESRI shape files served by NEFCS via their
ftp site.  Permitted names are BTS_Strata, EcoMon_Strata, Shellfish_Strata,
Shrimp_Shrimp, Statistical_Areas or All}

\item{dest}{fully qualified path to destination directory, by default ".".
If it doesn't exists then an attempt is made to create it.}

\item{ftp_url}{character ftp address, by default ftp://ftp.nefsc.noaa.gov/pub/gis}

\item{exts}{character vector of file extensions to download}
}
\value{
named logical vector - one for each file downloaded
}
\description{
Retreive one or more shape files from NMFS's NEFSC public GIS products
}

