% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{quadm}
\alias{quadm}
\title{Given a quadrisphere code, retrieve the multipliers for lat and lon}
\usage{
quadm(what)
}
\arguments{
\item{what}{one of "nw", "sw", "ne", "se"}
}
\value{
a 2 element vector of lat and lon multipliers
}
\description{
Given a quadrisphere code, retrieve the multipliers for lat and lon
}

