% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishbox.R
\name{show_fishboxes}
\alias{show_fishboxes}
\title{A toy example that shows the boxes within the One Degree Square '4269'}
\usage{
show_fishboxes(x = 4269)
}
\arguments{
\item{x}{one 4 digit code}
}
\description{
A toy example that shows the boxes within the One Degree Square '4269'
}

